Cast floating point number to qreal.
This makes it handle double/float differences appropriately. Otherwise
QtTest asserts when run on ARM.

+++ templates/tests/testinternationalization.cpp
@@ -805,7 +805,7 @@ void TestInternationalization::testFloats_data()
   QTest::addColumn<QString>("frFloatingPoint");
 
   QTest::newRow("float-01")
-    << 7.4
+    << qreal(7.4)
     << QString::fromLatin1("7.4")
     << QString::fromLatin1("7.40")
     << QString::fromLatin1("7.40")
